Added admin editing functions.
Added the ability to make admin on register. This works with no problems! Yay!

Added functionality to show who is admin, and default box checked if they are! Double yay!

Problems with changing person to/from admin by toggling the table's checkbox. The problem seems to be a conflict between current state and state of the database.  I'm not sure how to resolve it.