WIP book creation/database
The task is mostly done, WIP. Temporarily as proof of concept I have the user submitting the author name and book title. Later I will make sure it grabs author as the email address from getPeople.

I am also in the process of making it map a table of the book database but I haven't finished writing it, but I wanted to WIP commit.