Fixed delete book, added approval stamps, added functionality of approval button (not finished), rearranged InfoRegister.
Delete book was deleting only the books and not the pages. I totally forgot about this so the DB has like 300 pages for books that don't exist haha.

Approval button added and shows IF you are not the original book creator. This isn't perfect, but essentially forces two people to approve a translation.